requirements.txt
PyJWT
bcrypt
cryptography